restructure libs, bg writer again
