+++ src/lib.rs
@@ -1,5 +1,7 @@
 use pyo3::prelude::*;
 mod bamcoverage;
+mod compute;
+mod alignmentsieve;
 
 #[pymodule]
 fn hp(m: &Bound<'_, PyModule>) -> PyResult<()> {
